+++ ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: lrocigno <lrocigno@student.42sp.org>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/03 12:25:18 by lrocigno          #+#    #+#             */
-/*   Updated: 2021/05/11 17:13:00 by lrocigno         ###   ########.fr       */
+/*   Updated: 2021/10/28 11:43:33 by lrocigno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_intlen(intmax_t n)
+size_t	ft_intlen(int n)
 {
 	size_t		len;
