+++ ft_uintlen.c
@@ -6,13 +6,13 @@
 /*   By: lrocigno <lrocigno@student.42sp.org>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/30 21:41:59 by lrocigno          #+#    #+#             */
-/*   Updated: 2021/09/14 22:58:13 by lrocigno         ###   ########.fr       */
+/*   Updated: 2021/10/28 11:54:03 by lrocigno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_uintlen(uintmax_t n)
+size_t	ft_uintlen(unsigned int n)
 {
 	size_t		len;
